+++ app/src/main/java/com/infinity/omos/adapters/BindingAdapters.kt
@@ -2,7 +2,7 @@ package com.infinity.omos.adapters
 
 import android.view.View
 import androidx.databinding.BindingAdapter
-import com.infinity.omos.ui.onboarding.OmosFieldView
+import com.infinity.omos.ui.view.OmosFieldView
 import com.infinity.omos.ui.onboarding.ErrorField
 
 @BindingAdapter("isGone")
