+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -41,7 +41,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/iv_logo" />
 
-        <com.infinity.omos.ui.onboarding.OmosFieldView
+        <com.infinity.omos.ui.view.OmosFieldView
             android:id="@+id/ofv_email"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -65,7 +65,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toBottomOf="@id/ofv_email" />
 
-        <com.infinity.omos.ui.onboarding.OmosFieldView
+        <com.infinity.omos.ui.view.OmosFieldView
             android:id="@+id/ofs_email_auth_code"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
