+++ app/src/main/res/layout/fragment_login.xml
@@ -41,7 +41,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/iv_logo" />
 
-        <com.infinity.omos.ui.onboarding.OmosFieldView
+        <com.infinity.omos.ui.view.OmosFieldView
             android:id="@+id/ofv_email"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -53,7 +53,7 @@
             app:layout_constraintTop_toBottomOf="@id/tv_login"
             app:title="@string/email" />
 
-        <com.infinity.omos.ui.onboarding.OmosFieldView
+        <com.infinity.omos.ui.view.OmosFieldView
             android:id="@+id/ofv_password"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
